+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to
 - 🐛(frontend) show actual selection count in hard delete modal
 - 🐛(frontend) Responsive broken with long filters in search #659
 - 🐛(front) set size and variant on trash navigate modal #666
+- 🐛(frontend) fix uploads continuing after parent folder deletion
 
 ## [v0.16.0] - 2026-04-09